feat: improved text animations
